# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API including SVT and TV4 Play",
        "title": "Play API",
        "contact": {
            "name": "API Support",
            "url": "http://xxxx.xxx.xx",
            "email": "support@egeback.se"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.5"
    },
    "basePath": "/api/v1/",
    "paths": {
        "/ping": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common"
                ],
                "summary": "ping common",
                "responses": {
                    "200": {
                        "description": "pong",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/shows": {
            "get": {
                "description": "get shows",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shows"
                ],
                "summary": "List shows",
                "parameters": [
                    {
                        "type": "string",
                        "format": "bool",
                        "description": "pretty print show",
                        "name": "prettyPrint",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "bool",
                        "description": "show seasons",
                        "name": "showSeasons",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Show"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shows/{slug}": {
            "get": {
                "description": "get show by slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shows"
                ],
                "summary": "Show an show",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show ID",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "bool",
                        "description": "pretty print show",
                        "name": "prettyPrint",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Show"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/controllers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                },
                "more_info": {
                    "type": "string",
                    "example": "http://"
                }
            }
        },
        "models.Episode": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "longDescription": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "platform_specific": {
                    "type": "object"
                },
                "slug": {
                    "description": "SvtID           string    `json:\"svtId\" groups:\"api\"`\nVideoSvtID      string    `json:\"videoSvtId\" groups:\"api\"`",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "validFrom": {
                    "type": "string"
                },
                "validTo": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Variant"
                    }
                }
            }
        },
        "models.Season": {
            "type": "object",
            "properties": {
                "episodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Episode"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Show": {
            "type": "object",
            "properties": {
                "api_url": {
                    "type": "string",
                    "example": "http://adad.ad/se"
                },
                "decription": {
                    "type": "string",
                    "example": "Show about x"
                },
                "genre": {
                    "type": "string",
                    "example": "2019-12-22"
                },
                "imageUrl": {
                    "type": "string",
                    "example": "http://adad.ad/se"
                },
                "name": {
                    "type": "string",
                    "example": "Show Name"
                },
                "page_url": {
                    "type": "string",
                    "example": "http://adad.ad/se"
                },
                "seasons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Season"
                    }
                },
                "service": {
                    "type": "string"
                },
                "slug": {
                    "type": "string",
                    "example": "show_name"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2019-12-22"
                }
            }
        },
        "models.Variant": {
            "type": "object",
            "properties": {
                "platform_specific": {
                    "type": "object"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}